Use Object class for encoding Tuple
When building a Tuple where the types of each element are not the
same, use the Object class. This should be a rare case.

Exposed for me by the time-series example, where I'm representing the
date as Long but the tracking info as String.

Time-series example code is committed in
https://github.com/vedang/farstar