Add a decode function to the core
This simplifies the byte-array -> to Tuple/Subspace translation code.
It is the first step towards making all the parsing functions optional
again, since it you store your keys and values as the built-in
layers (i.e. Tuples, Subspaces or Directories) then the library can
auto translate them into and out of byte-arrays.

Note that if the decode function cannot decode your data, it will
simply return it as a byte-array. You (the user) has to convert it
successfully to data. A future commit will support a `:valfn` option
to pass in the desired conversion function.